(upd):command, move utils.

+++ lua/core/plug.lua
@@ -16,9 +16,7 @@ function Plug:load_plugins()
   local get_plugins_list = function()
     local list = {}
     local tmp = vim.split(fn.globpath(modules_dir, '*/plugins.lua'), '\n')
-    for _, f in ipairs(tmp) do
-      list[#list + 1] = f:sub(#modules_dir - 6, -1)
-    end
+    for _, f in ipairs(tmp) do list[#list + 1] = f:sub(#modules_dir - 6, -1) end
     return list
   end
 
@@ -42,17 +40,15 @@ function Plug:load_packer()
     display = {
       open_fn = function()
         return require('packer.util').float({border = 'single'})
-      end
+      end,
     },
-    disable_commands = true
+    disable_commands = true,
   })
   packer.reset()
   local use = packer.use
   self:load_plugins()
   use {"wbthomason/packer.nvim", opt = true}
-  for _, repo in ipairs(self.repos) do
-    use(repo)
-  end
+  for _, repo in ipairs(self.repos) do use(repo) end
 end
 
 function Plug:init_ensure_plugins()
@@ -60,11 +56,10 @@ function Plug:init_ensure_plugins()
   local state = uv.fs_stat(packer_dir)
   if not state then
     local cmd = "!git clone https://github.com/wbthomason/packer.nvim " ..
-                    packer_dir
+                  packer_dir
     api.nvim_command(cmd)
-    uv.fs_mkdir(data_dir .. 'lua', 511, function()
-      assert("make compile path dir faield")
-    end)
+    uv.fs_mkdir(data_dir .. 'lua', 511,
+      function() assert("make compile path dir faield") end)
     self:load_packer()
     packer.install()
   end
@@ -72,16 +67,12 @@ end
 
 local plugins = setmetatable({}, {
   __index = function(_, key)
-    if not packer then
-      Plug:load_packer()
-    end
+    if not packer then Plug:load_packer() end
     return packer[key]
-  end
+  end,
 })
 
-function plugins.ensure_plugins()
-  Plug:init_ensure_plugins()
-end
+function plugins.ensure_plugins() Plug:init_ensure_plugins() end
 
 function plugins.convert_compile_file()
   local lines = {}
@@ -92,9 +83,7 @@ function plugins.convert_compile_file()
     lnum = lnum + 1
     if lnum > 15 then
       lines[#lines + 1] = line .. '\n'
-      if line == 'END' then
-        break
-      end
+      if line == 'END' then break end
     end
   end
   table.remove(lines, #lines)
@@ -103,14 +92,10 @@ function plugins.convert_compile_file()
     os.execute('mkdir -p ' .. data_dir .. 'lua')
   end
 
-  if vim.fn.filereadable(compile_to_lua) == 1 then
-    os.remove(compile_to_lua)
-  end
+  if vim.fn.filereadable(compile_to_lua) == 1 then os.remove(compile_to_lua) end
 
   local file = io.open(compile_to_lua, "w")
-  for _, line in ipairs(lines) do
-    file:write(line)
-  end
+  for _, line in ipairs(lines) do file:write(line) end
   file:close()
 
   os.remove(packer_compiled)
@@ -134,17 +119,26 @@ function plugins.load_compile()
   if vim.fn.filereadable(compile_to_lua) == 1 then
     require('_compiled')
   else
-    -- assert(
-    --     'Missing packer compile file Run PackerCompile Or PackerInstall to fix')
+    assert(
+      'Missing packer compile file Run PackerCompile Or PackerInstall to fix')
     plugins.magic_compile()
   end
-  vim.cmd [[command! PlugCompile lua require('core.plug').magic_compile()]]
-  vim.cmd [[command! PlugInstall lua require('core.plug').install()]]
-  vim.cmd [[command! PlugUpdate lua require('core.plug').update()]]
-  vim.cmd [[command! PlugSync lua require('core.plug').sync()]]
-  vim.cmd [[command! PlugClean lua require('core.plug').clean()]]
-  vim.cmd [[autocmd User PackerComplete lua require('core.plug').magic_compile()]]
-  vim.cmd [[command! PlugStatus  lua require('core.plug').status()]]
+  r17.command {
+    "PlugCompile",
+    [[call v:lua.require('core.plug').magic_compile()]],
+  }
+  r17.command {"PlugInstall", [[lua require('core.plug').install()]]}
+  r17.command {"PlugSync", [[lua require('core.plug').sync()]]}
+  r17.command {"PlugClean", [[lua require('core.plug').clean()]]}
+  r17.command {"PlugUpdate", [[lua require('core.plug').update()]]}
+  r17.command {"PlugStatus", [[lua require('core.plug').status()]]}
+  r17.augroup("PackerComplete", {
+    {
+      events = {"User"},
+      targets = {"lua"},
+      command = "lua require('core.plug').magic_compile()",
+    },
+  })
 end
 
 return plugins
